// packages/api-contracts/contract-ts/src/models/FamilyMemberContext.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WishPool API
 * WishPool core API contract for mobile apps, parent web, admin web, and backend services. All write endpoints are command-oriented and should be called with an Idempotency-Key header. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FamilyMember } from './FamilyMember';
import {
    FamilyMemberFromJSON,
    FamilyMemberFromJSONTyped,
    FamilyMemberToJSON,
    FamilyMemberToJSONTyped,
} from './FamilyMember';
import type { Family } from './Family';
import {
    FamilyFromJSON,
    FamilyFromJSONTyped,
    FamilyToJSON,
    FamilyToJSONTyped,
} from './Family';

/**
 * 
 * @export
 * @interface FamilyMemberContext
 */
export interface FamilyMemberContext {
    /**
     * 
     * @type {Family}
     * @memberof FamilyMemberContext
     */
    family: Family;
    /**
     * 
     * @type {FamilyMember}
     * @memberof FamilyMemberContext
     */
    member: FamilyMember;
}

/**
 * Check if a given object implements the FamilyMemberContext interface.
 */
export function instanceOfFamilyMemberContext(value: object): value is FamilyMemberContext {
    if (!('family' in value) || value['family'] === undefined) return false;
    if (!('member' in value) || value['member'] === undefined) return false;
    return true;
}

export function FamilyMemberContextFromJSON(json: any): FamilyMemberContext {
    return FamilyMemberContextFromJSONTyped(json, false);
}

export function FamilyMemberContextFromJSONTyped(json: any, ignoreDiscriminator: boolean): FamilyMemberContext {
    if (json == null) {
        return json;
    }
    return {
        
        'family': FamilyFromJSON(json['family']),
        'member': FamilyMemberFromJSON(json['member']),
    };
}

export function FamilyMemberContextToJSON(json: any): FamilyMemberContext {
    return FamilyMemberContextToJSONTyped(json, false);
}

export function FamilyMemberContextToJSONTyped(value?: FamilyMemberContext | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'family': FamilyToJSON(value['family']),
        'member': FamilyMemberToJSON(value['member']),
    };
}