// packages/api-contracts/contract-ts/src/models/MeResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WishPool API
 * WishPool core API contract for mobile apps, parent web, admin web, and backend services. All write endpoints are command-oriented and should be called with an Idempotency-Key header. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { User } from './User';
import {
    UserFromJSON,
    UserFromJSONTyped,
    UserToJSON,
    UserToJSONTyped,
} from './User';
import type { FamilyMemberContext } from './FamilyMemberContext';
import {
    FamilyMemberContextFromJSON,
    FamilyMemberContextFromJSONTyped,
    FamilyMemberContextToJSON,
    FamilyMemberContextToJSONTyped,
} from './FamilyMemberContext';

/**
 * 
 * @export
 * @interface MeResponse
 */
export interface MeResponse {
    /**
     * 
     * @type {User}
     * @memberof MeResponse
     */
    user: User;
    /**
     * 
     * @type {Array<FamilyMemberContext>}
     * @memberof MeResponse
     */
    families: Array<FamilyMemberContext>;
}

/**
 * Check if a given object implements the MeResponse interface.
 */
export function instanceOfMeResponse(value: object): value is MeResponse {
    if (!('user' in value) || value['user'] === undefined) return false;
    if (!('families' in value) || value['families'] === undefined) return false;
    return true;
}

export function MeResponseFromJSON(json: any): MeResponse {
    return MeResponseFromJSONTyped(json, false);
}

export function MeResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): MeResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'user': UserFromJSON(json['user']),
        'families': ((json['families'] as Array<any>).map(FamilyMemberContextFromJSON)),
    };
}

export function MeResponseToJSON(json: any): MeResponse {
    return MeResponseToJSONTyped(json, false);
}

export function MeResponseToJSONTyped(value?: MeResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'user': UserToJSON(value['user']),
        'families': ((value['families'] as Array<any>).map(FamilyMemberContextToJSON)),
    };
}